CPU chip passed the test script. A day and a half's plodding hard work.
Bugs fixed:

a-bit being 0 means use A, a-bit being 1 means use M, I mixed the two up.

To match the ALU chart from left to right as provided in textbook, x input should be D and y input should be A or M. I swapped the two.

Also ALU's zr is j2 and ng is j1. I also misunderstood the matching--the ALU will only ever give for j1j2j3 a single 1, since a number that's greater than 0 then won't be equal to zero, et cetera. So the ALU's info output is exact whereas it just has to match any from the instruction. So if the instruction is 1 1 1 for the j1j2j3 bits then with the ALU being 1 0 0 it'd cause a jump. Fixed that.

Lastly a reminder: because A-register updates occur from the A-instruction address or from the ALU if a C-instruction is given, and that I guarantee no jumps if it's an A-instruction since it won't load the PC, I can always look at the C-instruction bit on line 60's Mux16 prior to the ALU because it won't affect anything if the bit is from an A-instruction due to the guards previously mentioned.